Move command line running and data file utils into core
