+++ src/main/kotlin/org/jetbrains/research/testspark/tools/llm/generation/grazie/GrazieRequestManager.kt
@@ -41,7 +41,7 @@ class GrazieRequestManager : RequestManager() {
                                 TestSparkBundle.message("tooLongPrompt"),
                                 project,
                             )
-                            sendResult = SendResult.TOOLONG
+                            sendResult = SendResult.TOO_LONG
                         }
 
                         else -> {
