+++ src/main/kotlin/org/jetbrains/research/testspark/tools/llm/generation/openai/OpenAIRequestManager.kt
@@ -55,7 +55,7 @@ class OpenAIRequestManager : RequestManager() {
                             TestSparkBundle.message("tooLongPrompt"),
                             project,
                         )
-                        sendResult = SendResult.TOOLONG
+                        sendResult = SendResult.TOO_LONG
                     }
 
                     HttpURLConnection.HTTP_UNAUTHORIZED -> {
